+++ test/openssl/test_ossl.rb
@@ -11,9 +11,6 @@ class OpenSSL::OSSL < OpenSSL::SSLTestCase
     assert_raise(ArgumentError) { OpenSSL.fixed_length_secure_compare("aaa", "aa") }
 
     assert OpenSSL.fixed_length_secure_compare("aaa", "aaa")
-    assert OpenSSL.fixed_length_secure_compare(
-      OpenSSL::Digest.digest('SHA256', "aaa"), OpenSSL::Digest::SHA256.digest("aaa")
-    )
 
     assert_raise(ArgumentError) { OpenSSL.fixed_length_secure_compare("aaa", "aaaa") }
     refute OpenSSL.fixed_length_secure_compare("aaa", "baa")